added AppBar component
