+++ client/src/components/AppBar.tsx
@@ -1,0 +1,17 @@
+import { Box } from "grommet";
+
+const AppBar = (props: any) => (
+    <Box
+      tag='header'
+      direction='row'
+      align='center'
+      justify='between'
+      background='brand'
+      pad={{ left: 'medium', right: 'small', vertical: 'small' }}
+      elevation='medium'
+      style={{ zIndex: '1' }}
+      {...props}
+    />
+  );
+
+export { AppBar };
